`svm_alloc` export (#170)

+++ crates/svm-runtime/src/vmcalls/calldata.rs
@@ -6,17 +6,17 @@ use wasmer_runtime::Ctx as WasmerCtx;
 pub fn calldata_ptr(ctx: &mut WasmerCtx) -> i32 {
     use_gas!("calldata_ptr", ctx);
 
-    calldata(ctx).0
+    calldata(ctx).0 as i32
 }
 
 pub fn calldata_len(ctx: &mut WasmerCtx) -> i32 {
     use_gas!("calldata_len", ctx);
 
-    calldata(ctx).1
+    calldata(ctx).1 as i32
 }
 
 #[inline]
-fn calldata(ctx: &mut WasmerCtx) -> (i32, i32) {
+fn calldata(ctx: &mut WasmerCtx) -> (usize, usize) {
     let svm_ctx = helpers::wasmer_data_svm(ctx.data);
 
     svm_ctx.get_calldata()
